dash-leaflet-app-1/requirements.txt
Dash==2.0.0
dash-leaflet==0.1.0
psycopg2-binary==2.9.1
pandas==1.3.3
numpy==1.21.2
plotly==5.1.0